Room_booking_serve/requirements.txt
<<<<<<< HEAD
asgiref==3.7.2
Django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
=======
langchain==0.2.10
langchain-core==0.2.22
langchain-openai==0.1.10
openai==1.30.5

vosk==0.3.44
sounddevice==0.5.5
scipy==1.11.4

pytest==8.2.2
python-dotenv==1.2.1
requests==2.32.5
PyYAML==6.0.3
>>>>>>> 2f3bcc1c3007826557e7a3123caad90bb3064cc5
